Chope down the module
